+++ Plugins/Bugsnag/Source/Bugsnag/Public/BugsnagApp.h
@@ -57,7 +57,7 @@ public:
 
 	// versionCode?: number;
 
-	virtual const TOptional<int64> GetVersionCode() const = 0;
+	virtual const TOptional<int32> GetVersionCode() const = 0;
 
-	virtual void SetVersionCode(const TOptional<int64>&) = 0;
+	virtual void SetVersionCode(const TOptional<int32>&) = 0;
 };
